AUTH refactor(auth): consolidate recovery types and schemas, remove deprecated file
