+++ .devcontainer/devcontainer.json
@@ -1,42 +1,42 @@
-// {
-//   "name": "Next.js Web (Docker)",
-//   "dockerComposeFile": ["../containers/docker-compose.yml"],
-//   "service": "frontend",
-//   "workspaceFolder": "/app",
-//   "shutdownAction": "stopCompose",
-//   "customizations": {
-//     "vscode": {
-//       "settings": {
-//         "editor.formatOnSave": true,
-//         "editor.defaultFormatter": "esbenp.prettier-vscode",
-//         "[typescriptreact]": {
-//           "editor.defaultFormatter": "esbenp.prettier-vscode"
-//         },
-//         "[javascriptreact]": {
-//           "editor.defaultFormatter": "esbenp.prettier-vscode"
-//         },
-//         "terminal.integrated.defaultProfile.linux": "bash"
-//       },
-//       "files.watcherExclude": {
-//         "**/.next/**": true,
-//         "**/node_modules/**": true,
-//         "**/.git/**": true
-//       },
+{
+  "name": "Next.js Web (Docker)",
+  "dockerComposeFile": ["../containers/docker-compose.yml"],
+  "service": "frontend",
+  "workspaceFolder": "/app",
+  "shutdownAction": "stopCompose",
+  "customizations": {
+    "vscode": {
+      "settings": {
+        "editor.formatOnSave": true,
+        "editor.defaultFormatter": "esbenp.prettier-vscode",
+        "[typescriptreact]": {
+          "editor.defaultFormatter": "esbenp.prettier-vscode"
+        },
+        "[javascriptreact]": {
+          "editor.defaultFormatter": "esbenp.prettier-vscode"
+        },
+        "terminal.integrated.defaultProfile.linux": "bash"
+      },
+      "files.watcherExclude": {
+        "**/.next/**": true,
+        "**/node_modules/**": true,
+        "**/.git/**": true
+      },
 
-//       "search.exclude": {
-//         "**/.next": true,
-//         "**/node_modules": true
-//       },
-//       "extensions": [
-//         "dbaeumer.vscode-eslint",
-//         "esbenp.prettier-vscode",
-//         "rhalaly.scope-to-this",
-//         "bradlc.vscode-tailwindcss",
-//         "Gruntfuggly.todo-tree"
-//       ]
-//     }
-//   }
-// }
+      "search.exclude": {
+        "**/.next": true,
+        "**/node_modules": true
+      },
+      "extensions": [
+        "dbaeumer.vscode-eslint",
+        "esbenp.prettier-vscode",
+        "rhalaly.scope-to-this",
+        "bradlc.vscode-tailwindcss",
+        "Gruntfuggly.todo-tree"
+      ]
+    }
+  }
+}
 // {
 //   "name": "Express.js backend (Docker)",
 //   "dockerComposeFile": ["../containers/docker-compose.yml"],
